# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Nifty100 Analytics API","description":"REST API for Nifty100 financial analytics","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/health":{"get":{"summary":"Health Check","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/companies":{"get":{"tags":["Companies"],"summary":"Get Companies","description":"Return all Nifty 100 companies.","operationId":"get_companies_api_v1_companies_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/companies/{ticker}":{"get":{"tags":["Companies"],"summary":"Get Company","description":"Return company profile information.","operationId":"get_company_api_v1_companies__ticker__get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/pl":{"get":{"tags":["Companies"],"summary":"Get Profit And Loss","description":"Return profit and loss history for a company.","operationId":"get_profit_and_loss_api_v1_companies__ticker__pl_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/bs":{"get":{"tags":["Companies"],"summary":"Get Balance Sheet","description":"Return balance sheet history for a company.","operationId":"get_balance_sheet_api_v1_companies__ticker__bs_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/cashflow":{"get":{"tags":["Companies"],"summary":"Get Cashflow","description":"Return cash-flow history for a company.","operationId":"get_cashflow_api_v1_companies__ticker__cashflow_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/ratios":{"get":{"tags":["Companies"],"summary":"Get Financial Ratios","description":"Return financial ratio history for a company.","operationId":"get_financial_ratios_api_v1_companies__ticker__ratios_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/tearsheet":{"get":{"tags":["Companies"],"summary":"Get Tearsheet","description":"Return the available company tearsheet document.","operationId":"get_tearsheet_api_v1_companies__ticker__tearsheet_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/screener":{"get":{"tags":["Screener"],"summary":"Screen Companies","description":"Screen Nifty 100 companies using latest available financial ratios.\n\nSupported filters:\n- min_roe\n- max_debt_to_equity\n- min_revenue_cagr\n- min_quality_score","operationId":"screen_companies_api_v1_screener_get","parameters":[{"name":"min_roe","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"description":"Minimum return on equity percentage","title":"Min Roe"},"description":"Minimum return on equity percentage"},{"name":"max_debt_to_equity","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"description":"Maximum debt-to-equity ratio","title":"Max Debt To Equity"},"description":"Maximum debt-to-equity ratio"},{"name":"min_revenue_cagr","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"description":"Minimum 5-year revenue CAGR percentage","title":"Min Revenue Cagr"},"description":"Minimum 5-year revenue CAGR percentage"},{"name":"min_quality_score","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"description":"Minimum composite quality score","title":"Min Quality Score"},"description":"Minimum composite quality score"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sectors":{"get":{"tags":["Sectors"],"summary":"Get Sectors","description":"Return all broad sectors with company counts.","operationId":"get_sectors_api_v1_sectors_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/sectors/{sector}/companies":{"get":{"tags":["Sectors"],"summary":"Get Sector Companies","description":"Return companies belonging to a specified broad sector.","operationId":"get_sector_companies_api_v1_sectors__sector__companies_get","parameters":[{"name":"sector","in":"path","required":true,"schema":{"type":"string","title":"Sector"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/peers/{group_name}":{"get":{"tags":["Peers"],"summary":"Get Peer Group","description":"Return companies belonging to a peer group.","operationId":"get_peer_group_api_v1_peers__group_name__get","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/peers/compare":{"get":{"tags":["Peers"],"summary":"Compare Company With Peers","description":"Compare a company with the other companies in its peer group.\n\nUses the latest available financial ratio year for each company.","operationId":"compare_company_with_peers_api_v1_companies__ticker__peers_compare_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/market-cap/{ticker}":{"get":{"tags":["Valuation"],"summary":"Get Market Cap","description":"Return the latest market-cap and valuation data for a company.","operationId":"get_market_cap_api_v1_market_cap__ticker__get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/portfolio/stats":{"get":{"tags":["Portfolio"],"summary":"Get Portfolio Stats","operationId":"get_portfolio_stats_api_v1_portfolio_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/companies/{ticker}/documents":{"get":{"tags":["Documents"],"summary":"Get Company Documents","operationId":"get_company_documents_api_v1_companies__ticker__documents_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}